fix: 分离日志与主类
