+++ nana7mi/__init__.py
@@ -1,120 +1,6 @@
-import asyncio
-import datetime
-import logging
-import os
-import sys
-from functools import wraps
-from importlib import import_module
-
-from apscheduler.schedulers.asyncio import AsyncIOScheduler
-
-class Logger:
-    # 日志
-    __logger = logging.getLogger('Nana7mi')
-    __logger.setLevel(logging.INFO)
-    handler = logging.StreamHandler()
-    handler.setFormatter(logging.Formatter("[%(asctime)s] [%(levelname)s] %(message)s", '%H:%M:%S'))
-    __logger.addHandler(handler)
-
-    def info(self, msg: str, id: str = ''):
-        self.__logger.info(f'[{id}] {msg}' if id else msg)
-
-    def error(self, msg: str, id: str = ''):
-        self.__logger.error(f'[{id}] {msg}' if id else msg)
-
-    def debug(self, msg: str, id: str = ''):
-        self.__logger.debug(f'[{id}] {msg}' if id else msg)
-
-log = Logger()
-
-from .adapters.basebot import BaseBot
-from .adapters.event import Message, MessageType, ParseLimit, limit, on_command
-
-class Nana7mi:
-    def __init__(self):
-        '基于 OneBot 的机器人框架'
-        # 适配器字典
-        self.bot_dict = dict()
-        # 插件集
-        self.response = list()
-        # 自动任务相关
-        self.sched = AsyncIOScheduler(timezone="Asia/Shanghai")
-        self.run_time = lambda seconds: datetime.datetime.now() + datetime.timedelta(seconds=seconds)
-
-    def setResponse(self, command = None, plt: ParseLimit = None):
-        def response_function(func):
-            @wraps(func)
-            @on_command(command)
-            @limit(plt)
-            async def wrapper(event):
-                return await func(event)
-            funcInfo = f'(command={command}'
-            if plt:
-                funcInfo += ''.join([f', {k}={v}' for k, v in plt.items() if v])
-            funcInfo += ')'
-            for wfunc, info in self.response:
-                if wfunc.__name__ == wrapper.__name__ and info == funcInfo:
-                    break
-            else:
-                self.response.append((wrapper, funcInfo))
-                log.info(f' » {wrapper.__name__}{funcInfo}')
-                return wrapper
-            return func
-        return response_function
-
-    def register_adapter(self, bot: BaseBot, name: str = ''):
-        # 安装适配器
-        if not name:
-            name = bot.__class__.__name__
-        assert isinstance(bot, BaseBot), f'注册 {name}: {bot} 时错误，非继承自基础机器人'
-        self.bot_dict[name] = bot.setParent(self)
-
-    def load_buildin_plugins(self):
-        log.info('内部模块已导入')
-        # 重启命令
-        @self.setResponse('/reboot', ParseLimit(white_user=[3099665076, 144115218677563300], group_both_user=True, guild_both_user=True))
-        async def reboot(event: Message):
-            event._receiver._closed = True
-            await event._receiver.aws.close_connection()
-
-        # 响应来自 cqbot 的位置命令
-        @self.setResponse(command='/here')
-        async def here(event: Message):
-            if event.message_type == MessageType.Guild:
-                return f'该频道ID：{event.guild_id}\n子频道ID：{event.channel_id}'
-
-        # 响应来自 cqbot 的回声命令
-        @self.setResponse(command='/echo')
-        async def echo(event: Message):
-            log.info(str(event), 'echo')
-            return event.content.replace('&#91;', '[').replace('&#93;', ']')
-
-        log.info('-------------')
-        return self
-
-    def load_plugins(self, folder: str):
-        sys.path.append(folder)
-        for root, dirs, files in os.walk(folder):
-            for file in files:
-                if not file.startswith('_') and file.endswith('.py'):
-                    try:
-                        log.info(f'{file} 已导入')
-                        import_module(file.replace('.py', ''))
-                        log.info('-------------')
-                    except Exception as e:
-                        log.error(f'{file} 加载错误：{e}')
-            break
-        return self
-
-    def run(self):
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-        self.sched.start()
-        pending = list()
-        for name, bot in self.bot_dict.items():
-            log.info(f'{name} 启动中', id='Nana7mi')
-            pending.append(bot.run(loop))
-        loop.run_until_complete(asyncio.wait(pending, return_when=asyncio.FIRST_COMPLETED))
+from .logger import log
+from .nana7mi import Nana7mi
+from .adapters.event import Message, MessageType, ParseLimit
 
 __THE_ONLY_ONE_BOT = Nana7mi()
